% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_extra_signal_threshold}
\alias{default_extra_signal_threshold}
\title{Return the default extra signal threshold}
\usage{
default_extra_signal_threshold()
}
\value{
A numeric value.
}
\description{
Returns the threshold used to flag a relevant third sequence.
}
