% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_heterozygote_threshold}
\alias{default_heterozygote_threshold}
\title{Return the default general allele balance threshold}
\usage{
default_heterozygote_threshold()
}
\value{
A numeric value.
}
\description{
Returns the default general threshold used to decide whether the second
most abundant sequence should be accepted as a second haplotype.
}
