% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_loci_order}
\alias{default_loci_order}
\title{Return the default iSNP loci order}
\usage{
default_loci_order()
}
\value{
A character vector with iSNP locus names.
}
\description{
Returns the default loci order used to create the wide TXT file for
Familias. The first column in the TXT is always \code{sample_id}; the loci
returned by this function are appended after it.
}
