% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_min_heterozygous_reads}
\alias{default_min_heterozygous_reads}
\title{Return the default minimum reads for heterozygous calls}
\usage{
default_min_heterozygous_reads()
}
\value{
A numeric value.
}
\description{
Returns the minimum number of reads required for the second sequence to be
accepted as the second allele in a heterozygous call.
}
