% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_min_homozygote_reads}
\alias{default_min_homozygote_reads}
\title{Return the default minimum reads for homozygous duplicated calls}
\usage{
default_min_homozygote_reads()
}
\value{
A numeric value.
}
\description{
Returns the minimum number of reads required to duplicate a single sequence
as a homozygous call.
}
